Made Z80 access safe
In the "Guide-01" document in the Rex Sabio manual dump, the risk of an
interrupt occurring during a Z80 write, and deasserting the Z80 bus
request, is mentioned. This is a risk in S3K and SCD, but only SCD
protects itself from it. Since that is also a risk with this sound
driver, I ported SCD's workaround.