Implemented better fix for not-updating-modulation bug
In Sonic 1 (and I imagine most SMPS 68k drivers), the modulation is not
updated or applied on the frame a new note starts. This only matters on
held notes.

Sonic 2 and Ristar both fixed this, but to my understanding they both do
it wrong. Ristar only fixes the modulation not being *applied*, not
updated. S2 makes it do both, but in such a hackish and overkill way.
Basically it run the 'update frequency' code twice in a row. This newer
way doesn't do that.