Use a new CycleSoundQueue
That old one was *very* wonky: backing-up registers it didn't need
to, and writing the new priority after the new sound had started
playing (potentially undoing a priority write that the sound handler
had performed).

I've also never been a fan of SMPS's habit of only playing one thing
per frame, and pushing lower-priority sounds out of the queue just
because something higher-priority happened to be in the queue on the
same frame (even though it has the 'do not store this priority' bit
set), so I've overhauled the priority system to play multiple things
per frame, stutter be damned.

If this is a problem for you, then maybe don't try to play four
different songs on the same frame, eh?