Change how the end of DAC samples is detected
Like Dual PCM, the DAC driver expects sample to mark the ends of
themselves with a 00 byte. This removes the need to allocate a whole
16-bit register to tracking the remaining number of iterations. It
also speed the code up significantly, and the spare register allows
for even further optimisations. The DAC driver now runs at around
17kHz, whereas previously it was around 15kHz.

I've added a small converter tool to the extras folder which converts
PCM samples to this format. I guess I'll have to extend with the
ability to convert DPCM samples as well, but that can wait for later.