Move the DAC table to the 68k
Making the table the Z80's responsibility meant having it use up a
chunk of the Z80's RAM, and it limited what the DAC driver could do:
for instance, what if the user wants to specify their own sample rate
at runtime?

I've also implemented a basic framework for implementing DAC
commands, but it's not finished yet.