Convert paths to relative
This modification seems to be required for compatibility with the new
custom AS builds by Flamewing. Unfortunate that there isn't a way to
maintain compatibility with both old and new versions.

I suppose if you're using this driver in an older project which
targets an older version of AS, then just revert this commit and
everything should hopefully work once again.