Identify this driver as a v2.8 prototype
'v2.7+' was a temporary name, but now there's a GitHub release with
that title, so it's not really suitable anymore.

Considering how heavily the driver has been modified since v2.7+, I
think it's safe to say that the new version will be v2.8 and not just
v2.7.1, so we can safely say that this driver is now a v2.8
prototype.