Prevent a potential freeze in `SMPS_Setup`
Fixes #7.

Until this commit, an interrupt could occur in the middle of
`SMPS_Setup`. This function is dependent on the Z80 bus being given
to the 68k, however the interrupt could execute code that returns it.
In that case, when the interrupt ends and the function resumes, it
will do so without the 68k having the Z80 bus, causing a whole bunch
of problems.

Sonic 1 and 2 aren't vulnerable to this by default, but elaborate
usage of this driver could be. Sonic 2's `SoundDriverLoad` actually
includes this exact fix, so it clearly isn't a far-fetched idea that
this issue could occur.