Move user files out of the 'Sonic-2-Clone-Driver-v2' directory.
Because of the root .gitignore, if a project included this driver as
a Git subrepository, then its music and SFX and such would never be
included in the Git history. This is stupid, and has bit me in the
arse in both my Hyper Sonic in Sonic 1 hack and my Sound Test
homebrew. It's dumb and it ends now.

I should really do the same for the PCM samples and PSG envelopes,
since users may want to add their own without forking the driver.